合并了EquipmentController
